Added search recent, error banner, shared prefs, update ui

+++ movies-client/src/main/java/com/enaz/movies/client/repository/MoviesRepository.kt
@@ -3,7 +3,10 @@ package com.enaz.movies.client.repository
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.enaz.movies.client.MoviesApiClient
+import com.enaz.movies.client.R
 import com.enaz.movies.client.model.MoviesResponse
+import com.enaz.movies.common.manager.SharedPreferencesManager
+import com.enaz.movies.common.manager.SharedPreferencesManagerImpl
 import com.enaz.movies.database.dao.MovieDao
 import com.enaz.movies.database.entity.MovieEntity
 import com.enaz.movies.util.clientModelToMovieEntity
@@ -14,22 +17,27 @@ import kotlin.coroutines.CoroutineContext
  * Created by eduardo.delito on 5/15/20.
  */
 interface MoviesRepository {
-//    suspend fun getMovies(search: String?): MoviesResponse
+    val recentSearch: LiveData<Pair<Int, String>?>
 
-    val errorMessage: LiveData<Int?>
+    val errorBanner: LiveData<Pair<Boolean, String>?>
 
     val loading: LiveData<Boolean>
 
-    fun searchMovies(search: String?)
+    fun searchMovies(search: String)
 
     fun getMovies(): LiveData<List<MovieEntity>>
 
-    fun deleteProducts()
+    fun getRecent()
+
+    fun deleteMovies()
+
+    fun resetBanner()
 }
 
 class MoviesRepositoryImpl(
     private val moviesApiClient: MoviesApiClient,
-    private val movieDao: MovieDao
+    private val movieDao: MovieDao,
+    private val sharedPreferencesManager: SharedPreferencesManager
 ) : MoviesRepository, CoroutineScope {
 
     override val coroutineContext: CoroutineContext
@@ -38,41 +46,79 @@ class MoviesRepositoryImpl(
     private val _loading = MutableLiveData<Boolean>()
     override val loading: LiveData<Boolean> get() = _loading
 
-    private val _errorMessage = MutableLiveData<Int?>()
-    override val errorMessage: LiveData<Int?> get() = _errorMessage
+    private val _recentSearch = MutableLiveData<Pair<Int, String>?>()
+    override val recentSearch: LiveData<Pair<Int, String>?> get() = _recentSearch
 
-//    override suspend fun getMovies(search: String?) = moviesApiClient.getMoviesResponse().getMovies(search, COUNTRY, MEDIA)
+    private val _errorBanner = MutableLiveData<Pair<Boolean, String>?>()
+    override val errorBanner: LiveData<Pair<Boolean, String>?> get() = _errorBanner
 
     override fun getMovies() = movieDao.getMovies()
 
-    override fun searchMovies(search: String?) {
+    override fun searchMovies(search: String) {
         _loading.postValue(true)
         launch {
-            insertMoviesBG(moviesApiClient.getMoviesResponse().getMovies(search, COUNTRY, MEDIA))
+            insertMoviesBG(
+                moviesApiClient.getMoviesResponse().getMovies(search, COUNTRY, MEDIA),
+                search
+            )
         }
     }
 
-    private suspend fun insertMoviesBG(moviesResponse : MoviesResponse) {
+    private suspend fun insertMoviesBG(moviesResponse: MoviesResponse, search: String) {
         withContext(Dispatchers.IO) {
             try {
                 _loading.postValue(false)
-                movieDao.deleteAll()
-                movieDao.insertMovies(moviesResponse.results.clientModelToMovieEntity())
+                val list = moviesResponse.results.clientModelToMovieEntity()
+                if (list.isNotEmpty()) {
+                    movieDao.deleteAll()
+                    movieDao.insertMovies(list)
+                    sharedPreferencesManager.lastSearch(
+                        SharedPreferencesManagerImpl.LAST_SEARCH,
+                        search
+                    )
+                    _recentSearch.postValue(null)
+                } else {
+                    _errorBanner.postValue(
+                        Pair(
+                            true,
+                            search
+                        )
+                    )
+                }
             } catch (e: Exception) {
                 _loading.postValue(false)
-                _errorMessage.postValue(null)
             }
         }
     }
 
-    override fun deleteProducts() {
-        launch { deleteProductsBG() }
+    override fun deleteMovies() {
+        launch { deleteMoviesBG() }
     }
 
-    private suspend fun deleteProductsBG() {
+    private suspend fun deleteMoviesBG() {
         withContext(Dispatchers.IO) {
             movieDao.deleteAll()
         }
+    }
+
+    override fun getRecent() {
+        val recent = sharedPreferencesManager.lastSearch(SharedPreferencesManagerImpl.LAST_SEARCH)
+        _recentSearch.postValue(
+            if (!recent.isNullOrEmpty())
+                Pair(
+                    R.string.recent_search,
+                    sharedPreferencesManager.lastSearch(SharedPreferencesManagerImpl.LAST_SEARCH)
+                ) else null
+        )
+    }
+
+    override fun resetBanner() {
+        _errorBanner.postValue(
+            Pair(
+                false,
+                SharedPreferencesManagerImpl.EMPTY_STRING
+            )
+        )
     }
 
     companion object {
